Refactor: 테스트용 Mocking Redis Config 변경 및 테스트 유틸 주석 추가
