+++ knock-backend/core/core-api/src/test/java/com/knock/ContextTest.java
@@ -1,6 +1,6 @@
 package com.knock;
 
-import com.knock.core.v1.TestRedisConfig;
+import com.knock.core.config.TestMockRedisConfig;
 import org.junit.jupiter.api.Tag;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
@@ -10,7 +10,7 @@ import org.springframework.test.context.TestConstructor;
 @Tag("context")
 @ActiveProfiles("test")
 @SpringBootTest
-@Import(TestRedisConfig.class)
+@Import(TestMockRedisConfig.class)
 @TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
 public abstract class ContextTest {
 }
