+++ knock-backend/core/core-api/src/test/java/com/knock/core/api/controller/v1/config/TestSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.knock.core.api.controller.v1;
+package com.knock.core.api.controller.v1.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
