+++ knock-backend/core/core-api/src/test/java/com/knock/core/config/TestMockRedisConfig.java
@@ -1,4 +1,4 @@
-package com.knock.core.v1;
+package com.knock.core.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -8,8 +8,11 @@ import org.springframework.data.redis.core.RedisTemplate;
 
 import static org.mockito.Mockito.mock;
 
+/**
+ * Redis Test Mocking Config
+ */
 @TestConfiguration
-public class TestRedisConfig {
+public class TestMockRedisConfig {
 
     @Bean
     @Primary
