+++ knock-backend/tests/api-docs/src/main/java/com/knock/test/api/RestDocsUtils.java
@@ -31,8 +31,7 @@ public class RestDocsUtils {
 			}
 
 			@Override
-			public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-					NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+			public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
 				return principal;
 			}
 		};
